// web/src/types/monthlyChange.ts
export type MonthlyChangeValue = number | null;

export type MonthlyChangeRow = {
  year: string;
  m01: MonthlyChangeValue;
  m02: MonthlyChangeValue;
  m03: MonthlyChangeValue;
  m04: MonthlyChangeValue;
  m05: MonthlyChangeValue;
  m06: MonthlyChangeValue;
  m07: MonthlyChangeValue;
  m08: MonthlyChangeValue;
  m09: MonthlyChangeValue;
  m10: MonthlyChangeValue;
  m11: MonthlyChangeValue;
  m12: MonthlyChangeValue;
};